+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
+export * from './controlPropKeys';
 export * from './defaults';
 export * from './types';
 export * from './utils';